+++ ${project.build.directory}/classes/application-prod.properties
@@ -1,0 +1,3 @@
+server.port=8090
+usr=@usr@
+pas=@pas@
