+++ pom.xml
@@ -13,6 +13,7 @@
 	<version>0.0.1-SNAPSHOT</version>
 	<name>examen</name>
 	<description>Demo project for Spring Boot</description>
+
 	<properties>
 		<java.version>11</java.version>
 	</properties>
@@ -27,6 +28,7 @@
 			<artifactId>lombok</artifactId>
 			<optional>true</optional>
 		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
@@ -47,23 +49,71 @@
 			<version>12.1.0.jre11-preview</version>
 		</dependency>
 
+		<!-- https://mvnrepository.com/artifact/io.springfox/springfox-boot-starter -->
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-boot-starter</artifactId>
+			<version>3.0.0</version>
+		</dependency>
+
+
+		<!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-security -->
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-security</artifactId>
+			<version>2.7.5</version>
+		</dependency>
+
+		<!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-data-jpa -->
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-jpa</artifactId>
+			<version>2.7.5</version>
+		</dependency>
+
 	</dependencies>
 
+	<!-- PERFILES -->
+	<profiles>
+
+		<!-- Perfil Desarrollo -->
+		<profile>
+			<id>dev</id>
+			<activation>
+				<activeByDefault>true</activeByDefault>
+			</activation>
+			<properties>
+				<spring.profieles.active>dev</spring.profieles.active>
+				<usr>${usr}</usr>
+				<pas>${pas}</pas>
+			</properties>
+		</profile>
+
+		<!-- Perfil Produccion -->
+		<profile>
+			<id>prod</id>
+			<properties>
+				<spring.profieles.active>prod</spring.profieles.active>
+			</properties>
+		</profile>
+
+	</profiles>
+
 	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-maven-plugin</artifactId>
-				<configuration>
-					<excludes>
-						<exclude>
-							<groupId>org.projectlombok</groupId>
-							<artifactId>lombok</artifactId>
-						</exclude>
-					</excludes>
-				</configuration>
-			</plugin>
-		</plugins>
+		<!--		<plugins>-->
+		<!--			<plugin>-->
+		<!--				<groupId>org.springframework.boot</groupId>-->
+		<!--				<artifactId>spring-boot-maven-plugin</artifactId>-->
+		<!--				<configuration>-->
+		<!--					<excludes>-->
+		<!--						<exclude>-->
+		<!--							<groupId>org.projectlombok</groupId>-->
+		<!--							<artifactId>lombok</artifactId>-->
+		<!--						</exclude>-->
+		<!--					</excludes>-->
+		<!--				</configuration>-->
+		<!--			</plugin>-->
+		<!--		</plugins>-->
 	</build>
 
 </project>
