--- src/main/java/com/example/entity/DanielDto.java
@@ -1,17 +1,0 @@
-package com.example.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class DanielDto {
-    private Integer id;
-    private String nombre1;
-    private String nombre2;
-    private String Aperllido1;
-    private String Aperllido2;
-    private String cedula;
-    private Integer edad;
-    private String sexo;
-}
