+++ src/main/java/com/example/model/entity/DanielDto.java
@@ -1,0 +1,17 @@
+package com.example.model.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class DanielDto {
+    private Integer id;
+    private String nombre1;
+    private String nombre2;
+    private String Aperllido1;
+    private String Aperllido2;
+    private String cedula;
+    private Integer edad;
+    private String sexo;
+}
