+++ src/main/resources/application-dev.properties
@@ -1,0 +1,3 @@
+server.port=8070
+usr=@usr@
+pas=@pas@
